apps/ai-server/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "나형진",email = "skgudwls@konkuk.ac.kr"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi (>=0.120.0,<0.120.2)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pillow (>=11.0.0,<12.0.0)",
    "python-multipart (>=0.0.9,<0.1.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pymupdf (>=1.24.10,<1.26.0)",
    "pydantic (>=2.9.2,<3.0.0)",
    "fer (>=22.5.1,<22.6.0)",
    "opencv-python (>=4.10.0,<4.13.0)",
    "matplotlib (>=3.9.2,<3.11.0)",
    "pandas (>=2.2.3,<2.4.0)",
    "moviepy (==1.0.3)",
    "tensorflow (>=2.17.0,<3.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-core (>=0.3.72,<0.4.0)",
    "langchain-openai (>=0.3.28,<0.4.0)",
    "groq (>=0.9.0,<1.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.2"
